+++ react/components/utils.tsx
@@ -2,7 +2,6 @@ import React from 'react'
 import { ISuggestionQueryResponseSearch } from '../utils/biggy-client'
 import { Item } from './Autocomplete/components/SuggestionSection/types'
 import { decodeUrlString } from '../utils/string-utils'
-import { getCookie, setCookie } from '../utils/dom-utils'
 
 /**
  * Given a query and a label: wrap the term in a bold span for styling
@@ -130,4 +129,31 @@ export const addTermToHistory = () => {
       prependSearchHistory(term)
     }
   }
+}
+
+function getCookie(name: string): string | null {
+  const regex = new RegExp(`(^|;)[ ]*${name}=([^;]*)`)
+  const match = regex.exec(document.cookie)
+
+  if (!match) {
+    return null
+  }
+
+  try {
+    return decodeURIComponent(match[2])
+  } catch {
+    return match[2]
+  }
+}
+function setCookie(key: string, value: string, ttl?: number, path = '/') {
+  let expires = ''
+
+  if (ttl && ttl > 0) {
+    const expirationDate = new Date()
+
+    expirationDate.setTime(new Date().getTime() + ttl)
+    expires = `expires=${expirationDate.toUTCString()};`
+  }
+
+  document.cookie = `${key}=${encodeURIComponent(value)};${expires}path=${path}`
 }
